Conflicts:
	ArduPlane/ArduPlane.pde
	ArduPlane/system.pde
	ArduPlane/test.pde
	libraries/APO/Navigator_Dcm.cpp
	libraries/AP_ADC/AP_ADC_ADS7844.cpp
	libraries/AP_ADC/AP_ADC_ADS7844.h
	libraries/AP_DCM/AP_DCM.h
	libraries/AP_IMU/AP_IMU_INS.cpp
	libraries/AP_IMU/AP_IMU_INS.h
	libraries/AP_InertialSensor/AP_InertialSensor_Stub.h
	libraries/GCS_MAVLink/include/v0.9/ardupilotmega/testsuite.h
	libraries/GCS_MAVLink/include/v0.9/ardupilotmega/version.h
	libraries/GCS_MAVLink/include/v0.9/common/common.h
	libraries/GCS_MAVLink/include/v0.9/common/version.h
	libraries/GCS_MAVLink/include/v1.0/ardupilotmega/ardupilotmega.h
	libraries/GCS_MAVLink/include/v1.0/ardupilotmega/testsuite.h
	libraries/GCS_MAVLink/include/v1.0/ardupilotmega/version.h
	libraries/GCS_MAVLink/include/v1.0/common/version.h
	libraries/GCS_MAVLink/message_definitions/ardupilotmega.xml
	libraries/GCS_MAVLink/message_definitions_v1.0/ardupilotmega.xml

+++ libraries/AP_GPS/AP_GPS_Auto.cpp
@@ -55,7 +55,7 @@ AP_GPS_Auto::read(void)
 {
     GPS		*gps;
     uint8_t		i;
-    unsigned long then;
+    uint32_t then;
 
     // Loop through possible baudrates trying to detect a GPS at one of them.
     //
@@ -96,10 +96,10 @@ AP_GPS_Auto::read(void)
 GPS *
 AP_GPS_Auto::_detect(void)
 {
-    unsigned long then;
-    int		fingerprint[4];
-    int		tries;
-    int     charcount;
+    uint32_t then;
+    uint8_t	fingerprint[4];
+    uint8_t	tries;
+    uint16_t charcount;
     GPS		*gps;
 
     //
